+++ src/Analyzers/CSharp/Analysis/AddOrRemoveTrailingCommaAnalyzer.cs
@@ -32,6 +32,9 @@ public sealed class AddOrRemoveTrailingCommaAnalyzer : BaseDiagnosticAnalyzer
 
         context.RegisterSyntaxNodeAction(f => AnalyzeEnumDeclaration(f), SyntaxKind.EnumDeclaration);
         context.RegisterSyntaxNodeAction(f => AnalyzeAnonymousObjectCreationExpression(f), SyntaxKind.AnonymousObjectCreationExpression);
+#if ROSLYN_4_7
+        context.RegisterSyntaxNodeAction(f => AnalyzeCollectionExpression(f), SyntaxKind.CollectionExpression);
+#endif
 
         context.RegisterSyntaxNodeAction(
             f => AnalyzeInitializerExpression(f),
@@ -168,6 +171,51 @@ public sealed class AddOrRemoveTrailingCommaAnalyzer : BaseDiagnosticAnalyzer
             }
         }
     }
+
+#if ROSLYN_4_7
+    private static void AnalyzeCollectionExpression(SyntaxNodeAnalysisContext context)
+    {
+        TrailingCommaStyle style = context.GetTrailingCommaStyle();
+
+        if (style == TrailingCommaStyle.None)
+            return;
+
+        var objectCreation = (CollectionExpressionSyntax)context.Node;
+
+        SeparatedSyntaxList<CollectionElementSyntax> elements = objectCreation.Elements;
+
+        if (!elements.Any())
+            return;
+
+        int count = elements.Count;
+        int separatorCount = elements.SeparatorCount;
+
+        if (count == separatorCount)
+        {
+            if (style == TrailingCommaStyle.Omit)
+            {
+                ReportRemove(context, elements.GetSeparator(count - 1));
+            }
+            else if (style == TrailingCommaStyle.OmitWhenSingleLine
+                && elements.IsSingleLine(cancellationToken: context.CancellationToken))
+            {
+                ReportRemove(context, elements.GetSeparator(count - 1));
+            }
+        }
+        else if (separatorCount == count - 1)
+        {
+            if (style == TrailingCommaStyle.Include)
+            {
+                ReportAdd(context, elements.Last());
+            }
+            else if (style == TrailingCommaStyle.OmitWhenSingleLine
+                && !elements.IsSingleLine(cancellationToken: context.CancellationToken))
+            {
+                ReportAdd(context, elements.Last());
+            }
+        }
+    }
+#endif
 
     private static void ReportAdd(SyntaxNodeAnalysisContext context, SyntaxNode lastNode)
     {